chore(coll): 컬렉션 좋아요 서비스 구조 변경
컨벤션대로 CQRS 패턴 따르도록 변경했고, 그에 따라 테스트명도 수정
